♻️ Move formatDuration to lib/datetime and bind it in useDateTime
useDateTime() now provides a locale-bound formatDuration so components
don't thread the locale themselves; the pure function moves into
lib/datetime/format alongside the other formatters. Pure call sites
(createOptionsContextValue, event types list) keep the explicit-locale
form.
