+++ apps/web/src/lib/datetime/client.tsx
@@ -7,6 +7,7 @@ import React from "react";
 import type { DateInput, DateTimePreset } from "@/lib/datetime/format";
 import {
   formatDateTime as baseFormatDateTime,
+  formatDuration as baseFormatDuration,
   formatRelativeTime,
 } from "@/lib/datetime/format";
 import { getLocaleDefaults, getWeekdayNames } from "@/lib/datetime/locales";
@@ -96,6 +97,7 @@ export function useDateTime() {
           showTimeZone: opts?.showTimeZone,
         }),
       toRelativeTime: (value: DateInput) => formatRelativeTime(value, locale),
+      formatDuration: (minutes: number) => baseFormatDuration(minutes, locale),
       weekdays: () => getWeekdayNames(locale),
     }),
     [locale, timeZone, timeFormat],
