+++ apps/web/src/utils/date-time-utils.ts
@@ -38,31 +38,6 @@ export interface ParsedTimeSlotOption {
 
 export type ParsedDateTimeOpton = ParsedDateOption | ParsedTimeSlotOption;
 
-type DurationFormatCtor = new (
-  locale: string | undefined,
-  options: { style: "narrow" },
-) => { format(duration: { hours?: number; minutes?: number }): string };
-
-export const formatDuration = (minutes: number, locale?: string) => {
-  const hours = Math.floor(minutes / 60);
-  const mins = minutes % 60;
-  const DurationFormat = (Intl as { DurationFormat?: DurationFormatCtor })
-    .DurationFormat;
-  if (DurationFormat) {
-    return new DurationFormat(locale, { style: "narrow" }).format({
-      hours,
-      minutes: mins,
-    });
-  }
-  if (hours && mins) {
-    return `${hours}h ${mins}m`;
-  }
-  if (hours) {
-    return `${hours}h`;
-  }
-  return `${mins}m`;
-};
-
 export const removeAllOptionsForDay = (
   options: DateTimeOption[],
   date: Date,
